+++ src/js/upcoming/helpers.js
@@ -1,0 +1,7 @@
+export function getRandomItem(array) {
+  return array[Math.floor(Math.random() * array.length)];
+}
+
+export function roundToTen(number) {
+  return Math.floor(number * 10) / 10;
+}
